requirements.txt
faiss-cpu
langchain
openai
streamlit
streamlit-chat